+++ adapters/Chartboost/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Chartboost iOS Mediation Adapter Changelog
 
+#### Next version
+- Adapter now automatically forwards GDPR consent found inside `IABTCF_AddtlConsent` to the Chartboost SDK if GDPR applies.
+
 #### [Version 9.10.1.0](https://dl.google.com/googleadmobadssdk/mediation/ios/chartboost/ChartboostAdapter-9.10.1.0.zip)
 - Verified compatibility with Chartboost SDK version 9.10.1.